feat: Added additional tag to Cost Codes for MATs on School and Trust Cost Comparison
