+++ front-end-components/index.html
@@ -99,9 +99,10 @@
       data-id="02387916"
       data-phases='["Secondary","Primary","Post-16"]'
     ></div>-->
-    <!--<div id="compare-your-costs" data-type="school" data-id="150203" data-suppress-negative-or-zero="true"></div>-->
+    <div id="compare-your-costs" data-type="school" data-id="990250" data-suppress-negative-or-zero="true" 
+    data-cost-code-map="{&quot;teaching staff&quot;:&quot;BAE010&quot;,&quot;supply teaching staff&quot;:&quot;BAE020&quot;,&quot;agency supply teaching staff&quot;:&quot;BAE240&quot;,&quot;education support staff&quot;:&quot;BAE030&quot;,&quot;educational consultancy&quot;:&quot;BAE230&quot;,&quot;administrative and clerical staff&quot;:&quot;BAE040&quot;,&quot;auditor costs&quot;:&quot;BAE260&quot;,&quot;other staff&quot;:&quot;BAE070&quot;,&quot;professional services (non-curriculum)&quot;:&quot;BAE300&quot;,&quot;examination fees&quot;:&quot;BAE220&quot;,&quot;learning resources (not ICT equipment)&quot;:&quot;BAE200&quot;,&quot;ict learning resources&quot;:&quot;BAE210&quot;,&quot;cleaning and caretaking&quot;:&quot;BAE130&quot;,&quot;maintenance of premises&quot;:&quot;BAE120&quot;,&quot;other occupation costs&quot;:&quot;BAE180&quot;,&quot;premises staff&quot;:&quot;BAE050&quot;,&quot;energy&quot;:&quot;BAE150&quot;,&quot;water and sewerage&quot;:&quot;BAE140&quot;,&quot;administrative supplies (non-educational)&quot;:&quot;BAE280&quot;,&quot;catering staff&quot;:&quot;BAE060&quot;,&quot;catering supplies&quot;:&quot;BAE250&quot;,&quot;direct revenue financing&quot;:&quot;BAE290&quot;,&quot;grounds maintenance&quot;:&quot;BAE320&quot;,&quot;indirect employee expenses&quot;:&quot;BAE080&quot;,&quot;interest charges for loan and bank&quot;:&quot;BAE320&quot;,&quot;other insurance premiums&quot;:&quot;BAE270&quot;,&quot;private Finance Initiative (PFI) charges&quot;:&quot;BAE310&quot;,&quot;rent and rates&quot;:&quot;BAE160&quot;,&quot;special facilities&quot;:&quot;BAE190&quot;,&quot;staff development and training&quot;:&quot;BAE090&quot;,&quot;staff-related insurance&quot;:&quot;BAE110&quot;,&quot;supply teacher insurance&quot;:&quot;BAE100&quot;}" data-is-mat="true"></div>
     <!--<div id="historic-data" data-type="school" data-id="150963"></div>-->
-    <div id="historic-data-2" data-type="school" data-id="150963" data-phase="Primary" data-finance-type="Maintained"></div>
+    <!--<div id="historic-data-2" data-type="school" data-id="150963" data-phase="Primary" data-finance-type="Maintained"></div>-->
     <!--<div
       id="compare-your-costs"
       data-type="local-authority"
