+++ front-end-components/src/views/compare-your-costs/types.tsx
@@ -1,9 +1,10 @@
 import { CompareYourCostsProps } from "./partials/accordion-sections/types";
 
 export type CompareYourCostsViewProps = CompareYourCostsProps & {
+  costCodeMap?: Record<string, string>;
   customDataId: string | undefined;
   dispatchEventType?: string;
   phases: string[] | null;
   suppressNegativeOrZero: boolean;
-  costCodeMap: Record<string, string> | null;
+  tags?: string[];
 };
