+++ web/src/Web.App/Views/TrustComparison/Index.cshtml
@@ -39,7 +39,8 @@
          data-id="@Model.CompanyNumber"
          data-phases="@Model.Phases.ToJson(Formatting.None)"
          data-dispatch-event-type="@eventType"
-         data-cost-code-map="@Model.CostCodeMap.ToJson(Formatting.None)">
+         data-cost-code-map="@Model.CostCodeMap.ToJson(Formatting.None)"
+         data-is-mat="@Model.IsMat.ToString().ToLower()">
     </div>
 }
